Rename trilead test package to credentials
