+++ src/test/java/org/jenkinsci/plugins/gitclient/credentials/CredentialsProviderImplTest.java
@@ -1,4 +1,4 @@
-package org.jenkinsci.plugins.gitclient.trilead;
+package org.jenkinsci.plugins.gitclient.credentials;
 
 import static org.junit.Assert.*;
