+++ src/test/java/org/jenkinsci/plugins/gitclient/credentials/StandardUsernamePasswordCredentialsImpl.java
@@ -1,4 +1,4 @@
-package org.jenkinsci.plugins.gitclient.trilead;
+package org.jenkinsci.plugins.gitclient.credentials;
 
 import com.cloudbees.plugins.credentials.CredentialsDescriptor;
 import com.cloudbees.plugins.credentials.CredentialsScope;
